+++ Source/NSCell.m
@@ -502,7 +502,7 @@ static NSColor *dtxtCol;
    id="NSCellAttribute">NSCellAttribute</ref></p>
    <p>See Also: -setCellAttribute:to:</p>
  */
-- (int) cellAttribute: (NSCellAttribute)aParameter
+- (NSInteger) cellAttribute: (NSCellAttribute)aParameter
 {
   switch (aParameter)
     {
@@ -551,7 +551,7 @@ static NSColor *dtxtCol;
 /**<p>TODO</p>
  *<p>See Also: -cellAttribute:</p>
  */
-- (void) setCellAttribute: (NSCellAttribute)aParameter  to: (int)value
+- (void) setCellAttribute: (NSCellAttribute)aParameter to: (NSInteger)value
 {
   switch (aParameter)
     {
@@ -1172,9 +1172,9 @@ static NSColor *dtxtCol;
 
 /**<p>TODO Explain</p>
  */
-- (int) sendActionOn: (int)mask
+- (NSInteger) sendActionOn: (NSInteger)mask
 {
-  unsigned int previousMask = _action_mask;
+  NSUInteger previousMask = _action_mask;
 
   _action_mask = mask;
 
@@ -1239,15 +1239,15 @@ static NSColor *dtxtCol;
  * Formatting Data
  */
 - (void) setFloatingPointFormat: (BOOL)autoRange
-                           left: (unsigned int)leftDigits
-                          right: (unsigned int)rightDigits
+                           left: (NSUInteger)leftDigits
+                          right: (NSUInteger)rightDigits
 {
   NSNumberFormatter *formatter = [[NSNumberFormatter alloc] init];
   NSMutableString *format = [[NSMutableString alloc] init];
 
   if (autoRange)
     {
-      unsigned fieldWidth = leftDigits + rightDigits + 1;
+      NSUInteger fieldWidth = leftDigits + rightDigits + 1;
 
       // FIXME: this does not fully match the documentation.
       while (fieldWidth--)
@@ -1286,14 +1286,14 @@ static NSColor *dtxtCol;
 
 /**<p> TODO</p>
  */
-- (int) entryType
+- (NSInteger) entryType
 {
   return _cell.entry_type;
 }
 
 /** <p>TODO</p>
  */
-- (void) setEntryType: (int)aType
+- (void) setEntryType: (NSInteger)aType
 {
   [self setType: NSTextCellType];
   // TODO: This should select a suitable formatter
@@ -1384,7 +1384,7 @@ static NSColor *dtxtCol;
 
   if (r.length > 0)
     {
-      unsigned int location = r.location;
+      NSUInteger location = r.location;
       
       
       [self setTitle: [[aString substringToIndex: location] 
@@ -1579,7 +1579,7 @@ static NSColor *dtxtCol;
 /**<p>Returns the mouse flags. This flags are usally sets in 
    the -trackMouse:inRect:ofView:untilMouseUp: method</p>
  */
-- (int) mouseDownFlags
+- (NSInteger) mouseDownFlags
 { 
   return _mouse_down_flags;
 }
@@ -1627,7 +1627,7 @@ static NSColor *dtxtCol;
        untilMouseUp: (BOOL)flag
 {
   NSApplication *theApp = [NSApplication sharedApplication];
-  unsigned event_mask = NSLeftMouseDownMask | NSLeftMouseUpMask
+  NSUInteger event_mask = NSLeftMouseDownMask | NSLeftMouseUpMask
     | NSMouseMovedMask | NSLeftMouseDraggedMask | NSOtherMouseDraggedMask
     | NSRightMouseDraggedMask;
   NSPoint location = [theEvent locationInWindow];
@@ -2293,8 +2293,8 @@ static NSColor *dtxtCol;
                   inView: (NSView*)controlView
                   editor: (NSText*)textObject
                 delegate: (id)anObject
-                   start: (int)selStart
-                  length: (int)selLength
+                   start: (NSInteger)selStart
+                  length: (NSInteger)selLength
 {
   if (!controlView || !textObject || (_cell.type != NSTextCellType))
     return;
@@ -2526,7 +2526,7 @@ static NSColor *dtxtCol;
       if ([aDecoder containsValueForKey: @"NSCellFlags"])
         {
           unsigned long cFlags;
-          unsigned long mask = 0;
+          NSUInteger mask = 0;
           cFlags = [aDecoder decodeIntForKey: @"NSCellFlags"];
           
           [self setFocusRingType: (cFlags & 0x3)];
@@ -2979,7 +2979,7 @@ static NSColor *dtxtCol;
     }
 }
 
-- (BOOL) _sendsActionOn:(int)eventTypeMask
+- (BOOL) _sendsActionOn:(NSUInteger)eventTypeMask
 {
   return (_action_mask & eventTypeMask);
 }
